+++ studio/components/interfaces/Account/index.ts
@@ -1,0 +1,4 @@
+import AccessTokenList from './AccessTokenList'
+import NewAccessTokenButton from './NewAccessTokenButton'
+
+export { AccessTokenList, NewAccessTokenButton }
